Xbox-Live-status-notifier.py
Coded in Python so that it can be used on all systems. This particular bot is designed to use gmail, but you can exchange out the information for any email server you're currently using. Utilizes Beautiful Soup (in an ugly way) to parse out the Xbox live status page for changes to a certain phrase, meaning that something has gone wrong with the servers. It's not the prettiest script, but it gets the job done.

Any questions, comments, feel free to message me.