Big Update
Okay, so I basically rewrote a lot of it. Now the script is designed to do this:
1. Check if XBL is up, if it is, nothing happens
2. If XBL is down, it records the time it went down, parses out why it's down (this only looks at critical functions), and sends an email with that info in it.
3. It will keep analyzing the page until things are back up, it'll send a second email that says when it's over and how long it was down for.

Also included are error codes, which are nice. Since I'm running mine on a Windows server, I'll also include the batch script that I use to execute the task in task scheduler, which is designed to check if the script is running, and if it's not, to start it up again every five minutes. In this way, if the script receives an error for whatever reason (for example, the website is not accessible), it'll automatically close itself and restart five minutes later.